ci: fix Clippy + host-test failures on main after #43 merge
Two pre-existing red checks plus one regression introduced by #43 had to
be cleared so the pipeline goes green again.

Clippy (-D warnings):
* agc-core/src/programs/p29.rs — `Met(1_000_00)` had inconsistent digit
  grouping (from #20 P29); now `Met(100_000)`.
* agc-sim/src/runtime.rs:550 and agc-test/tests/uplink_scenarios.rs
  (three sites) cast `T4RUPT_PERIOD_CS as u32`, but the constant is
  already `u32` (mine, MS-U1/MS-U5).

Host tests:
* `tc_e7d_summary_rust_pipeline` checked the lunar_return Rust pipeline
  output against a reference summary captured during MS-E7d. MS-E6b
  refined the entry guidance (PREDANG F1/F2, GLIMITER) and the new miss
  distance / peak g / elapsed time no longer fit the old fixture's
  ±50 km / ±0.5 g / ±30 s tolerances. Regenerated both
  direct_leo_summary.json and lunar_return_summary.json by running the
  existing `regenerate_summary_fixtures` test (entry_e2e_vagc.rs:1023,
  which exists exactly for this case).

615 host tests passing after fix; clippy clean with the exact CI
invocation.

+++ agc-sim/src/runtime.rs
@@ -547,7 +547,7 @@ mod tests {
 
         // No further work — advance past one full T4 period; nothing must
         // change because the FIFO is empty.
-        state.time = Met(T4RUPT_PERIOD_CS as u32);
+        state.time = Met(T4RUPT_PERIOD_CS);
         t4.tick(&mut state, &mut hw);
         assert!(matches!(state.vn.phase, VnPhase::P27Address { .. }));
     }
